# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Social network API",
        "description": "API for social network site.",
        "version": "1.0"
    },
    "paths": {
...
        "/api/posts": {
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Update Post by id",
                "operationId": "UpdatePostById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID of Post",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    },
                    {
                        "name": "header",
                        "in": "query",
                        "description": "Header of the edit Post",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "header"
                    },
                    {
                        "name": "content",
                        "in": "query",
                        "description": "Content of the edit Post",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "content"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "post": {
                                            "properties": {
                                                "id": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "header": {
                                                    "type": "string",
                                                    "example": "title"
                                                },
                                                "content": {
                                                    "type": "string",
                                                    "example": "content"
                                                },
                                                "user_id": {
                                                    "type": "number",
                                                    "example": "1"
                                                },
                                                "imgage_src": {
                                                    "type": "string",
                                                    "example": "/storage/GxrusnxMd4oADTTz00e8aK8rVxuyvRuXIcx3oRca.png"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "No post found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Create Posts",
                "operationId": "NewPost",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image_src": {
                                        "description": "Upload image value",
                                        "type": "file",
                                        "format": "binary"
                                    },
                                    "header": {
                                        "description": "Add post header",
                                        "type": "string"
                                    },
                                    "content": {
                                        "description": "Add content post",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "header": {
                                            "type": "string",
                                            "example": "title"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "content"
                                        },
                                        "user_id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "imgage_src": {
                                            "type": "string",
                                            "example": "/storage/GxrusnxMd4oADTTz00e8aK8rVxuyvRuXIcx3oRca.png"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get Post by id",
                "operationId": "GetPostById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Post",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "post": {
                                            "properties": {
                                                "id": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "header": {
                                                    "type": "string",
                                                    "example": "title"
                                                },
                                                "content": {
                                                    "type": "string",
                                                    "example": "content"
                                                },
                                                "user_id": {
                                                    "type": "number",
                                                    "example": "1"
                                                },
                                                "imgage_src": {
                                                    "type": "string",
                                                    "example": "/storage/GxrusnxMd4oADTTz00e8aK8rVxuyvRuXIcx3oRca.png"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "No post found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Delete Post by id",
                "operationId": "DeletePostById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Post",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "post": {
                                            "properties": {
                                                "id": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "header": {
                                                    "type": "string",
                                                    "example": "title"
                                                },
                                                "content": {
                                                    "type": "string",
                                                    "example": "content"
                                                },
                                                "user_id": {
                                                    "type": "number",
                                                    "example": "1"
                                                },
                                                "imgage_src": {
                                                    "type": "string",
                                                    "example": "/storage/GxrusnxMd4oADTTz00e8aK8rVxuyvRuXIcx3oRca.png"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "No post found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Laravel passport apiKey security.",
                "in": "header",
                "scheme": "https",
                "name": "Authorization",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://laravel/oauth/authorize",
                        "tokenUrl": "http://laravel/oauth/token",
                        "refreshUrl": "http://laravel/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}